notes_ai/requirements.txt
streamlit==1.31.0
langchain==0.1.10
langchain-groq==0.0.1
google-cloud-vision==3.5.0
pytesseract==0.3.10
opencv-python==4.9.0.80
Pillow==10.2.0
reportlab==4.0.9
requests==2.31.0
duckduckgo-search==4.1.1
python-dotenv==1.0.0